replace homepage with search page
